app-2-python-docs-assistant/pyproject.toml
[project]
name = "python-docs-assistant"
version = "0.1.0"
description = "Advanced RAG app for Python documentation with hybrid search"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.2",
    "langchain-cohere>=0.3",
    "langchain-experimental>=0.3",
    "qdrant-client>=1.12",
    "rank-bm25>=0.2",
    "gradio>=5",
    "python-dotenv>=1.0",
    "requests>=2.31",
]
